+++ src/components/Language_dev.js
@@ -1,9 +1,7 @@
 import React, {Fragment} from 'react';
 import PropTypes from 'prop-types';
 import ValidLanguageList from '../utils/validLanguageList';
-import ValidNativeLanguageList from '../utils/validNativeLanguageList';
 import language from '../utils/language';
-import {getLocale, setLocale} from '../utils/locale';
 import List from '@material-ui/core/List';
 import ListSubheader from '@material-ui/core/List';
 import {withStyles} from '@material-ui/core/styles';
@@ -57,14 +55,6 @@ const styles = (theme) => ({
 			borderRadius: '0px',
 			marginBottom: '91px'
 		}
-	},
-	providedByInReach: {
-		display: 'flex',
-		fontFamily: 'arial',
-		fontSize: '11px',
-		color: '#666',
-		whiteSpace: 'nowrap',
-		padding: theme.spacing(2)
 	},
 	poweredByGoogle: {
 		display: 'flex',
@@ -149,20 +139,14 @@ class LangMenuItem extends React.Component {
 	}
 
 	handleSelectLang() {
-		this.props.handleSelectLang(
-			this.props.langCode,
-			this.props.langName,
-			this.props.provider
-		);
+		this.props.handleSelectLang(this.props.langCode, this.props.langName);
 	}
-
 	render() {
 		return (
 			<AsylumConnectDropdownListItem
 				data-test-id="nav-button-language-item"
 				button
 				onClick={this.handleSelectLang}
-				provider={this.props.provider}
 			>
 				{this.props.langName}
 			</AsylumConnectDropdownListItem>
@@ -175,10 +159,8 @@ class Language extends React.Component {
 		super();
 		this.state = {
 			open: false,
-			selectedLang: 'English',
 			langsList: ValidLanguageList.all(),
-			langsNativeList: ValidNativeLanguageList.all(),
-			provider: ''
+			selectedLang: 'English'
 		};
 		this.handleClick = this.handleClick.bind(this);
 		this.handleSelect = this.handleSelect.bind(this);
@@ -186,28 +168,10 @@ class Language extends React.Component {
 		this.handleRequestCloseAfterSelect =
 			this.handleRequestCloseAfterSelect.bind(this);
 		this.generateLanguageItems = this.generateLanguageItems.bind(this);
-		this.generateNativeLanguageItems =
-			this.generateNativeLanguageItems.bind(this);
 		this.generateLanguageList = this.generateLanguageList.bind(this);
 		this.generateLabelWithIcon = this.generateLabelWithIcon.bind(this);
 		this.handleOnFilterChange = this.handleOnFilterChange.bind(this);
 		this.handleOnFilterBarClick = this.handleOnFilterBarClick.bind(this);
-	}
-
-	generateNativeLanguageItems() {
-		return (
-			<Fragment>
-				{this.state.langsNativeList.map((lang, index) => (
-					<LangMenuItem
-						key={100 + index}
-						langName={lang.local}
-						langCode={lang['1']}
-						handleSelectLang={this.handleRequestCloseAfterSelect}
-						provider={ValidNativeLanguageList.provider}
-					/>
-				))}
-			</Fragment>
-		);
 	}
 
 	generateLanguageItems() {
@@ -215,11 +179,10 @@ class Language extends React.Component {
 			<Fragment>
 				{this.state.langsList.map((lang, index) => (
 					<LangMenuItem
-						key={200 + index}
+						key={index}
 						langName={lang.local}
 						langCode={lang['1']}
 						handleSelectLang={this.handleRequestCloseAfterSelect}
-						provider="gt"
 					/>
 				))}
 			</Fragment>
@@ -236,16 +199,6 @@ class Language extends React.Component {
 				].join(' ')}
 				spacing={3}
 			>
-				<ListSubheader className={this.props.classes.providedByInReach}>
-					<FormattedMessage
-						id="language.inreach-attribution"
-						defaultMessage="Provided by InReach"
-						description="Dropdown label describing language translator"
-					>
-						{(providedBy) => <span>{providedBy}</span>}
-					</FormattedMessage>
-				</ListSubheader>
-				{this.generateNativeLanguageItems()}
 				<div className={this.props.classes.filterInputBar}>
 					<Filter
 						className={this.props.classes.filterFormControl}
@@ -257,8 +210,7 @@ class Language extends React.Component {
 				<ListSubheader className={this.props.classes.poweredByGoogle}>
 					<FormattedMessage
 						id="language.google-attribution"
-						defaultMessage="Powered by Google Translate"
-						description="Dropdown label describing language translator"
+						defaultMessage="Powered by"
 					>
 						{(poweredBy) => <span>{poweredBy}</span>}
 					</FormattedMessage>
@@ -278,7 +230,6 @@ class Language extends React.Component {
 						<FormattedMessage
 							id="language.dropdown-translate"
 							defaultMessage="Powered by"
-							description="Dropdown label describing language translator"
 						>
 							{(translate) => (
 								<span className={this.props.classes.blackTranslateColor}>
@@ -307,72 +258,31 @@ class Language extends React.Component {
 	}
 
 	handleClick(event) {
-		this.setState(
-			{open: !this.state.open},
-			{selectedLang: language.getLanguage()}
-		);
+		this.setState({open: !this.state.open});
 	}
 
-	handleSelect(langCode, langName, provider) {
+	handleSelect(langCode, langName) {
 		if (typeof this.props.onSelect === 'function') {
-			this.props.onSelect(langCode, langName, provider);
-			this.setState({
-				selectedLanguage: langName
-			});
+			this.props.onSelect(langCode, langName);
 		}
 	}
 
 	handleOnFilterChange(e) {
 		const filteredList = ValidLanguageList.filteredLanguageList(e.target.value);
-		const filteredNativeList = ValidNativeLanguageList.filteredLanguageList(
-			e.target.value
-		);
 		this.setState({
-			langsList: filteredList,
-			langsNativeList: filteredNativeList
+			langsList: filteredList
 		});
 	}
 	handleOnFilterBarClick(e) {
 		e.stopPropagation();
 	}
 
-	handleRequestCloseAfterSelect(langCode, langName, provider) {
-		this.setState({open: false, selectedLang: langName, provider: provider});
-		if ((langCode === 'en' || langCode === 'es') && provider === 'inreach') {
-			this.setState({open: false, selectedLang: langName, provider: provider});
-			//clear location.hash
-			var uri = window.location.toString();
-			if (uri.indexOf('#') > 0) {
-				var clean_uri = uri.substring(0, uri.indexOf('#'));
-				window.history.replaceState({}, document.title, clean_uri);
-			}
-			//also clear googltrans cookie
-			document.cookie = 'googtrans=; path=/;Max-Age=0;';
-		} else {
-			//use google translate
-			window.location.hash = '#googtrans(' + langCode + ')';
-		}
+	handleRequestCloseAfterSelect(langCode, langName) {
+		this.setState({open: false, selectedLang: langName});
+		window.location.hash = '#googtrans(' + langCode + ')';
 		language.setLanguage(langName);
-		language.setLanguageCode(langCode);
-		provider
-			? language.setLanguageProvider(provider)
-			: language.removeLanguageProvider();
+		//window.localStorage.setItem('lang', langName);
 		this.handleSelect(langCode, langName);
-
-		if (langCode === 'es' && getLocale() === 'en_MX') {
-			setLocale('es_MX');
-		}
-		if (langCode === 'es' && getLocale() === 'en_US') {
-			setLocale('es_US');
-		}
-
-		if (langCode === 'en' && getLocale() === 'es_MX') {
-			setLocale('en_MX');
-		}
-		if (langCode === 'en' && getLocale() === 'es_US') {
-			setLocale('en_US');
-		}
-
 		if (this.props.autoReload) {
 			window.location.reload();
 		}
@@ -391,17 +301,14 @@ class Language extends React.Component {
 	componentWillMount() {
 		var currentLang = language.getLanguage(); //window.localStorage.getItem('lang') ? window.localStorage.getItem('lang') : 'English';
 		if (window.location.hash.length !== 0) {
-			let langCode = language.getLanguageCode();
-			currentLang =
-				ValidLanguageList.byCode(langCode) ||
-				ValidNativeLanguageList.byCode(langCode);
+			let langCode = window.location.hash
+				.substring(window.location.hash.indexOf('(') + 1)
+				.slice(0, -1)
+				.toLowerCase();
+			currentLang = ValidLanguageList.byCode(langCode);
 		}
 		this.setState({selectedLang: currentLang});
 		this.handleSelect(ValidLanguageList.codeByName(currentLang), currentLang);
-		this.handleSelect(
-			ValidNativeLanguageList.codeByName(currentLang),
-			currentLang
-		);
 		if (currentLang === 'English') {
 			document.cookie =
 				'googtrans=;path=/;expires=Thu, 01 Jan 1970 00:00:01 GMT;';
@@ -484,7 +391,6 @@ class Language extends React.Component {
 						<FormattedMessage
 							id="language.dropdown-select-language"
 							defaultMessage="Select Language"
-							description="Dropdown label to prompting to select a language"
 						>
 							{(selectLanguage) => (
 								<Typography className={classes.textCenter} variant="h3">
