+++ src/components/LocaleForm.js
@@ -6,6 +6,8 @@ import {withStyles} from '@material-ui/core/styles';
 
 import AsylumConnectButton from './AsylumConnectButton';
 import Language from './Language';
+import language from '../utils/language';
+
 import LocaleSelector from './LocaleSelector';
 import withWidth from './withWidth';
 import {getLocale} from '../utils/locale';
@@ -91,8 +93,8 @@ class LocaleForm extends React.Component {
 			reload: false,
 			selectedLanguage: false,
 			selectedLanguageName: false,
-			/*selectedLocale: false,
-      selectedLocaleName: false,*/
+			selectedLocale: false,
+			selectedLocaleName: false,
 			startingLang: this.getStartingLanguage()
 		};
 
@@ -118,13 +120,28 @@ class LocaleForm extends React.Component {
 	handleNextClick(ev) {
 		if (this.state.selectedLocale) {
 			this.props.changeLocale(this.state.selectedLocale);
-
-			//will need this once catalog is fully translasted to spanish
-			/* if(this.state.selectedLocale === 'en_MX' && this.state.selectedLanguage === 'es'){
+			//show app code in spanish if langCode is 'es' and locale is MX or US
+			if (
+				this.state.selectedLocale === 'en_MX' &&
+				this.state.selectedLanguage === 'es'
+			) {
 				this.props.changeLocale('es_MX');
-			}else {
-				this.props.changeLocale(this.state.selectedLocale);
-			} */
+			} else if (
+				this.state.selectedLocale === 'en_US' &&
+				this.state.selectedLanguage === 'es'
+			) {
+				this.props.changeLocale('es_US');
+			} else if (
+				this.state.selectedLocale === 'es_US' &&
+				this.state.selectedLanguage === 'en'
+			) {
+				this.props.changeLocale('en_US');
+			} else if (
+				this.state.selectedLocale === 'es_MX' &&
+				this.state.selectedLanguage === 'en'
+			) {
+				this.props.changeLocale('en_MX');
+			}
 		}
 
 		if (typeof this.props.onLocaleSelect === 'function') {
@@ -227,6 +244,7 @@ class LocaleForm extends React.Component {
 							listContainerClass={listContainerClass}
 							onSelect={this.handleSelectLanguage}
 							triggerReload={this.state.reload}
+							selectedLanguage={this.state.selectedLanguageName}
 						/>
 					</Grid>
 					<Grid item xs={12} md={6}>
