+++ src/components/LocaleForm_dev.js
@@ -6,8 +6,6 @@ import {withStyles} from '@material-ui/core/styles';
 
 import AsylumConnectButton from './AsylumConnectButton';
 import Language from './Language';
-import language from '../utils/language';
-
 import LocaleSelector from './LocaleSelector';
 import withWidth from './withWidth';
 import {getLocale} from '../utils/locale';
@@ -93,8 +91,8 @@ class LocaleForm extends React.Component {
 			reload: false,
 			selectedLanguage: false,
 			selectedLanguageName: false,
-			selectedLocale: false,
-			selectedLocaleName: false,
+			/*selectedLocale: false,
+      selectedLocaleName: false,*/
 			startingLang: this.getStartingLanguage()
 		};
 
@@ -120,28 +118,13 @@ class LocaleForm extends React.Component {
 	handleNextClick(ev) {
 		if (this.state.selectedLocale) {
 			this.props.changeLocale(this.state.selectedLocale);
-			//show app code in spanish if langCode is 'es' and locale is MX or US
-			if (
-				this.state.selectedLocale === 'en_MX' &&
-				this.state.selectedLanguage === 'es'
-			) {
+
+			//will need this once catalog is fully translasted to spanish
+			/* if(this.state.selectedLocale === 'en_MX' && this.state.selectedLanguage === 'es'){
 				this.props.changeLocale('es_MX');
-			} else if (
-				this.state.selectedLocale === 'en_US' &&
-				this.state.selectedLanguage === 'es'
-			) {
-				this.props.changeLocale('es_US');
-			} else if (
-				this.state.selectedLocale === 'es_US' &&
-				this.state.selectedLanguage === 'en'
-			) {
-				this.props.changeLocale('en_US');
-			} else if (
-				this.state.selectedLocale === 'es_MX' &&
-				this.state.selectedLanguage === 'en'
-			) {
-				this.props.changeLocale('en_MX');
-			}
+			}else {
+				this.props.changeLocale(this.state.selectedLocale);
+			} */
 		}
 
 		if (typeof this.props.onLocaleSelect === 'function') {
@@ -244,7 +227,6 @@ class LocaleForm extends React.Component {
 							listContainerClass={listContainerClass}
 							onSelect={this.handleSelectLanguage}
 							triggerReload={this.state.reload}
-							selectedLanguage={this.state.selectedLanguageName}
 						/>
 					</Grid>
 					<Grid item xs={12} md={6}>
